FEAT(deck): Rejigger components and styles
I added a `has-footer` class to `SlideViewLive` so that it's at least _theoretically_ possible that the footer might get positioned properly, if it has contents.

Because all the CSS is specified so !@#@%& specifically, with almost every single style being prefixed with `#main`, I can't figure out how to reposition the footer so that it's inside the viewport rather than below the bottom of it. So the footer may have to hang out in the wrong place until the whole .scss file is refactored, per Corey's recommendation.